refactor: to get ready for dockerization
